+++ ReleaseNotes.md
@@ -30,6 +30,8 @@
 * TGR-938: XML-Messages with UTF-8 content can now be transmitted without alteration.
 * TGR-1266: Invalid configurations for servers are not reported in console or workflow UI but abort the test env mgr
   silently.
+* TGR-1262: Step "TGR pause test run execution with message {string}" and step "TGR pause test run execution
+  with message {tigerResolvedString} and message in case of error {tigerResolvedString}" will now correctly distinguished
 
 # Release 2.3.2